requirements.txt
fastapi
uvicorn
pydantic
pandas
numpy
yfinance
scikit-learn
xgboost
shap
matplotlib
torch
langchain-text-splitters
langchain-community
langchain-core
langchain-huggingface
langchain-ollama
python-dotenv
faiss-cpu
sec-api
sentence-transformers